requirements.txt
requests==2.22.0
pathvalidate==2.2.0
python-dotenv==0.10.3
beautifulsoup4==4.7.1
lxml==4.5.0